requirements.txt
Django==1.3.1
Fabric==1.4.0
PIL==1.1.7
South==0.7.3
boto==2.2.2
django-appconf==0.5
django-ckeditor==0.0.9
django-compressor==1.1.2
django-debug-toolbar==0.9.4
django-ses==0.4.0
pycrypto==2.5
ssh==1.7.13